+++ corbeans-spring-boot-bnms-starter/src/main/kotlin/com/github/manosbatsis/corbeans/spring/boot/corda/bnms/web/CorbeansBmnsMemberPathFragmentController.kt
@@ -25,8 +25,14 @@ import com.r3.businessnetworks.membership.states.MembershipState
 import io.swagger.annotations.ApiOperation
 import io.swagger.annotations.ApiParam
 import org.slf4j.LoggerFactory
-import org.springframework.web.bind.annotation.*
-import java.util.*
+import org.springframework.web.bind.annotation.GetMapping
+import org.springframework.web.bind.annotation.PathVariable
+import org.springframework.web.bind.annotation.PostMapping
+import org.springframework.web.bind.annotation.PutMapping
+import org.springframework.web.bind.annotation.RequestBody
+import org.springframework.web.bind.annotation.RequestMapping
+import org.springframework.web.bind.annotation.RequestParam
+import java.util.Optional
 
 /**
  *  Exposes BNMS Member methods as endpoints.
@@ -78,6 +84,9 @@ open class CorbeansBmnsMemberPathFragmentController : CorbeansBmnsMemberBaseCont
             @ApiParam(value = "The BNO party name")
             @RequestParam(required = true)
             bno: String,
+            @ApiParam(value = "The network ID")
+            @RequestParam(required = false)
+            networkId: Optional<String>,
             @ApiParam(value = "Whether to force a refresh.")
             @RequestParam(required = false, defaultValue = "false")
             forceRefresh: Boolean,
@@ -85,7 +94,7 @@ open class CorbeansBmnsMemberPathFragmentController : CorbeansBmnsMemberBaseCont
             @RequestParam(required = false, defaultValue = "true")
             filterOutMissingFromNetworkMap: Boolean
     ): List<MembershipState<*>> =
-            super.listMemberships(nodeName, bno, forceRefresh, filterOutMissingFromNetworkMap)
+            super.listMemberships(nodeName, bno, networkId, forceRefresh, filterOutMissingFromNetworkMap)